+++ src/commands/index.ts
@@ -1,3 +1,4 @@
+export * from './Headline';
 export * from './RevisionText';
 export * from './SeoOptimization';
 export * from './TextEdits';