+++ src/extension.ts
@@ -1,10 +1,17 @@
 import { RevisionText } from './commands/RevisionText';
 import * as vscode from 'vscode';
-import { SeoOptimization, TextEdits } from './commands';
+import { headline, SeoOptimization, TextEdits } from './commands';
+import { SparkupContentProvider } from './provider';
 
 
 
 export function activate(context: vscode.ExtensionContext) {
+
+	new SparkupContentProvider(context).init();
+
+	context.subscriptions.push(
+		vscode.commands.registerCommand('vscode-sparkup.headline', headline.generate)
+	);
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-sparkup.fixSpelling', TextEdits.fixSpelling)
@@ -12,6 +19,10 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-sparkup.biasFreeLanguage', TextEdits.biasFreeLanguage)
+	);
+
+	context.subscriptions.push(
+		vscode.commands.registerCommand('vscode-sparkup.simplify', TextEdits.simplify)
 	);
 
 	context.subscriptions.push(