+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './getSelectedText';
 export * from './setSelectedText';
+export * from './sleep';