+++ src/utils/sleep.ts
@@ -1,0 +1,9 @@
+
+
+export const sleep = (time: number) => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve(true);
+    }, time);
+  });
+}